requirements.txt
playwright>=1.48,<2.0
# Minimum set for Sheets API
gspread==6.1.2
google-api-python-client==2.139.0
oauth2client==4.1.3
python-dotenv==1.0.1
